refactor:  固化本地 RAG 架构：384-d 维度锚定、全链路 Long ID 映射及 Native-First 逻辑切换。

+++ app/src/main/java/com/example/powerai/data/retriever/HttpAnnRetriever.kt
@@ -9,8 +9,8 @@ import javax.inject.Inject
 class HttpAnnRetriever @Inject constructor(
     private val api: AnnApiService
 ) : AnnRetriever {
-    override suspend fun search(query: String, k: Int): List<Int> {
+    override suspend fun search(query: String, k: Int): List<Long> {
         val resp = api.search(AnnSearchRequest(query, k))
-        return resp.results.map { it.id }
+        return resp.results.map { it.id.toLong() }
     }
 }
