+++ app/src/main/java/com/example/powerai/domain/retriever/AnnRetriever.kt
@@ -8,5 +8,5 @@ interface AnnRetriever {
      * Search for top-k nearest neighbor indices for a given query string.
      * Returns a list of integer indices corresponding to metadata positions in the embedding store.
      */
-    suspend fun search(query: String, k: Int = 10): List<Int>
+        suspend fun search(query: String, k: Int = 10): List<Long>
 }
